+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added number example for situnitx
 - Added switch `--githubpublish` to enable generating `README.md` for repositories hosted at <https://github.com/latextemplates>.
 - Added support for [minted](https://ctan.org/pkg/minted)
+- `.editorconfig` is now generated
 
 ### Changed